requirements.txt
streamlit
google-generativeai
fpdf
pydeck
